test(RoomView): move observableDatabase helper out of __tests__
